requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
export==0.2.1
ffmpeg==1.4
ffmpy==0.5.0
filelock==3.16.1
flatbuffers==24.12.23
fsspec==2024.12.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.69.0
h11==0.14.0
h5py==3.12.1
idna==3.10
Jinja2==3.1.5
keras==3.8.0
libclang==18.1.1
llvmlite==0.43.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.6.0
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
opt_einsum==3.4.0
optree==0.13.1
outcome==1.3.0.post0
packaging==24.2
pocketsphinx==5.0.4
protobuf==5.29.3
pycparser==2.22
pydub==0.25.1
Pygments==2.19.1
PySocks==1.7.1
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
SpeechRecognition==3.14.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tiktoken==0.8.0
torch==2.5.1
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
